requirements.txt
gradio==4.24.0
pandas==2.1.4
matplotlib==3.8.2
numpy==1.26.3
pydantic==2.10.6
pydantic-ai-slim[openai]==0.0.37
requests==2.32.3
python-dotenv==1.0.0
httpx==0.28.1 